requirements.txt
greenlet==3.0.1
pyTelegramBotAPI==4.14.0
pytz==2023.3.post1
schedule==1.2.1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
firebase-admin==6.5.0
psycopg2==2.9.9
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
numpy==2.0.1
pillow==10.4.0
proglog==0.1.10
requests==2.32.3
tqdm==4.66.4
urllib3==2.2.2
